[add]: CustomInventoryを生成するクラスを追加
